fix - domain names in console and final output
